requirements.txt
streamlit
crewai
crewai_tools
langchain_community
ChatOpenAI